add CharitiesHandler for GET and POST requests
